# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Partial_Users_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"points": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"LoginRequest": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			},
			"BookInterface": {
				"properties": {
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"coverImage": {
						"type": "string",
						"nullable": true
					},
					"writer": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"tags",
					"point",
					"coverImage",
					"writer",
					"title",
					"id"
				],
				"type": "object"
			},
			"Partial_BookInterface_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"writer": {
						"type": "string"
					},
					"coverImage": {
						"type": "string",
						"nullable": true
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"PurchaseRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"bookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"bookId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CancelRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bookstore-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {},
										"user": {
											"properties": {
												"points": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"lastName": {
													"type": "string"
												},
												"firstName": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"points",
												"password",
												"lastName",
												"firstName",
												"email",
												"deletedAt",
												"updatedAt",
												"createdAt",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"token",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Users_"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {},
										"user": {
											"properties": {
												"points": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"lastName": {
													"type": "string"
												},
												"firstName": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"points",
												"password",
												"lastName",
												"firstName",
												"email",
												"deletedAt",
												"updatedAt",
												"createdAt",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"token",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/books": {
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookInterface"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BookInterface"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_BookInterface_"
							}
						}
					}
				}
			}
		},
		"/books/search": {
			"get": {
				"operationId": "SearchBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BookInterface"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetBookById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BookInterface"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/BookInterface"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_BookInterface_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/orders/purchase": {
			"post": {
				"operationId": "PurchaseBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PurchaseRequest"
							}
						}
					}
				}
			}
		},
		"/orders/cancel/{orderId}": {
			"post": {
				"operationId": "CancelOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"book": {
													"properties": {
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"tags": {
															"items": {
																"type": "string"
															},
															"type": "array"
														},
														"point": {
															"type": "number",
															"format": "double"
														},
														"coverImage": {
															"type": "string"
														},
														"writer": {
															"type": "string"
														},
														"title": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"deletedAt",
														"updatedAt",
														"createdAt",
														"tags",
														"point",
														"coverImage",
														"writer",
														"title",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"book"
											],
											"type": "object"
										},
										{
											"properties": {
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"status": {
													"type": "string"
												},
												"bookId": {
													"type": "number",
													"format": "double"
												},
												"userId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"deletedAt",
												"updatedAt",
												"createdAt",
												"status",
												"bookId",
												"userId",
												"id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CancelRequest"
							}
						}
					}
				}
			}
		},
		"/orders/user/{userId}": {
			"get": {
				"operationId": "GetOrdersByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"deletedAt": {
												"type": "string",
												"format": "date-time"
											},
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"status": {
												"type": "string"
											},
											"bookId": {
												"type": "number",
												"format": "double"
											},
											"userId": {
												"type": "number",
												"format": "double"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"deletedAt",
											"updatedAt",
											"createdAt",
											"status",
											"bookId",
											"userId",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}